% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.rc.R
\name{data.frame.rc}
\alias{data.frame.rc}
\title{data.frame Creation, [R,C] defined}
\usage{
data.frame.rc(values, nrow, ncol = length(values),
  column_names = names(values))
}
\arguments{
\item{values}{Type: list or vector to repeat. The values inside the data.frame.}

\item{nrow}{Type: numeric. The number of rows in the data.frame to create.}

\item{ncol}{Type: numeric. The number of columns in the data.frame to create. Defaults to \code{length(values)}.}

\item{column_names}{Type: character. The names of the columns in the data.frame to create. Defaults to \code{names(values)}.}
}
\value{
A data.frame with [nrow, ncol] strucure specified, along with column names (defaults to \code{V1, V2, V3...})
}
\description{
\code{data.frame.rc} creates a \code{data.frame} using the [R,C] format like `matrix` (arguments: `nrow`, `ncol`).
}
\details{
Please see vignette for demos: \code{vignette("data.frame.rc", package = "Laurae2")} or \code{help_me("data.frame.rc")}.
}
\examples{

# Create default named [5, 5] data.frame
data.frame.rc(0, nrow = 5, ncol = 5)

# Create named [5, 5] data.frame
data.frame.rc(0, nrow = 5, ncol = 5,
              column_names = c("car", "minicooper", "truck", "porsche", "speedo"))

# Create named [5, 5] data.frame with values
data.frame.rc(list(1:5,
                   c(TRUE, FALSE, TRUE, FALSE, TRUE),
                   5:1,
                   c("has", "not", "has", "not", "has"),
                   c(150L, 200L, 250L, 300L, 350L)),
              nrow = 5, ncol = 5)

# Create named [5, 5] data.frame with named values
data.frame.rc(list(1:5,
                   c(TRUE, FALSE, TRUE, FALSE, TRUE),
                   5:1,
                   c("has", "not", "has", "not", "has"),
                   c(150L, 200L, 250L, 300L, 350L)),
              nrow = 5, ncol = 5,
              column_names = c("car", "minicooper", "truck", "porsche", "speedo"))

# Create named [5, 5] data.frame with values
data.frame.rc(list(car = 1:5,
                   minicooper = c(TRUE, FALSE, TRUE, FALSE, TRUE),
                   truck = 5:1,
                   porsche = c("has", "not", "has", "not", "has"),
                   speedo = c(150L, 200L, 250L, 300L, 350L)),
              nrow = 5, ncol = 5)
}
