% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_me.R
\name{help_me}
\alias{help_me}
\title{Open Laurae2 package Vignette}
\usage{
help_me(func_name = NULL, package = "Laurae2")
}
\arguments{
\item{func_name}{Type: character. A function name from the corresponding \code{package}.}

\item{package}{Type: character. The package the vignette must be looked for.}
}
\value{
Nothing.
}
\description{
\code{help_me} opens the correspdoning vignette to the function you are asking for.
}
\examples{
\dontrun{
help_me("data.table.rc")
}
}
