% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.bayesian.best.R
\name{optimize.bayesian.best}
\alias{optimize.bayesian.best}
\title{Bayesian Optimization, Best Result Set}
\usage{
optimize.bayesian.best(optimized)
}
\arguments{
\item{optimized}{Type: List returned from \code{optimize.bayesian}. The object to look for the best result.}
}
\value{
A list where \code{x} is the list of the best parameter set, where \code{y} is the loss value at the \code{x} parameters, and where \code{iter} is the best iteration number.
}
\description{
\code{optimize.bayesian.best} gets the best result issued from \code{optimize.bayesian}.
}
\details{
Please see vignette for demos: \code{vignette("optimize.bayesian", package = "Laurae2")} or \code{help_me("optimize.bayesian")}.
}
