% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.bayesian.R
\name{optimize.bayesian}
\alias{optimize.bayesian}
\title{Bayesian Optimization}
\usage{
optimize.bayesian(loss_func, param_set, seed = 1, maximize = FALSE,
  absurd_value = ifelse(maximize == TRUE, -Inf, Inf), exp_design = NULL,
  initialization = nrow(exp_design), max_evaluations = 50,
  time_budget = NULL, verbose = TRUE)
}
\arguments{
\item{loss_func}{Type: function. The loss function. Takes as input a list of elements.}

\item{param_set}{Type: ParamHelpers::makeParamSet. The parameters to optimize. Check out ParamHelpers::makeParamSet for more information.}

\item{seed}{Type: numeric. Seed for random number generation. Defaults to \code{1}.}

\item{maximize}{Type: logical. Whether to minimize (\code{FALSE}) or maximize (\code{TRUE}) the loss function. Defaults to \code{FALSE}.}

\item{absurd_value}{Type: numeric. An absurd value when an error or invalid value is provided by the loss function which is the reverse of the best possible loss value. Defaults to \code{ifelse(maximize == TRUE, -Inf, Inf)}.}

\item{exp_design}{Type: matrix. The default matrix used to initialize the optimizer. Defaults to \code{NULL}.}

\item{initialization}{Type: numeric. The number of \code{loss_func} initializations to provide a start point for the bayesian optimizer. Defaults to \code{nrow(exp_design)}.}

\item{max_evaluations}{Type: numeric. The number of times the loss function can be evaluated, including the initialization part. Defaults to \code{50}.}

\item{time_budget}{Type: numeric. The maximum time spent in the loss function globally. The optimization hands down the results to the user when the time budget is exhausted or the maximum number of evaluations is reached. Defaults to \code{NULL}.}

\item{verbose}{Type: logical. Whether to print debug messages. Defaults to \code{TRUE}.}
}
\description{
\code{optimize.bayesian} performs bayesian optimization on a given loss function and parameters. This is a wrapper to \code{mlrMBO} optimizer for quick and fast optimization.
}
\details{
Please see vignette for demos: \code{vignette("optimize.bayesian", package = "Laurae2")} or \code{help_me("optimize.bayesian")}.
}
