% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.bayesian.all.R
\name{optimize.bayesian.all}
\alias{optimize.bayesian.all}
\title{Bayesian Optimization, All Results Set}
\usage{
optimize.bayesian.all(optimized)
}
\arguments{
\item{optimized}{Type: List returned from \code{optimize.bayesian}. The object to look for all the results.}
}
\value{
A data.frame containing all the results set from the bayesian optimization. A column y refers to the loss function value for the according parameters on the same row.
}
\description{
\code{optimize.bayesian.all} gets all the results issued from \code{optimize.bayesian}.
}
\details{
Please see vignette for demos: \code{vignette("optimize.bayesian", package = "Laurae2")} or \code{help_me("optimize.bayesian")}.
}
